Moved mongo db files to its own mongo folder inside /db
