+++ classes/activity-command.js
@@ -2,7 +2,7 @@ const PermissionCommand = require('./permission-command');
 const discordServices = require('../discord-services');
 const Activity = require('./activity');
 const { Message } = require('discord.js');
-const BotGuild = require('../db/BotGuild');
+const BotGuild = require('../db/mongo/BotGuild');
 const BotGuildModel = require('./bot-guild');
 
 /**
