+++ classes/permission-command.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const { Command, CommandoMessage } = require('discord.js-commando');
-const BotGuild = require('../db/BotGuild');
+const BotGuild = require('../db/mongo/BotGuild');
 const BotGuildModel = require('./bot-guild');
 const discordServices = require('../discord-services');
